backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
psycopg[binary]==3.1.18
python-multipart==0.0.6
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
huggingface-hub==0.20.3
PyPDF2==3.0.1
python-docx==1.1.0
